Split native token burning and melting, add destroy examples
