+++ src/account/operations/transaction/high_level/burning_melting/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+pub(crate) mod burn_native_token;
 pub(crate) mod burn_nft;
 pub(crate) mod destroy_alias;
 pub(crate) mod destroy_foundry;
